refactor: 폴더 구조 반영 (#80)
