+++ src/pages/home/major-company-section/major-company-section.tsx
